moved tests and added non-top-level context
